requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --allow-unsafe requirements.in
#
amqp==5.2.0
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via starlette
billiard==4.2.0
    # via celery
celery[redis]==5.4.0
    # via -r requirements.in
certifi==2024.6.2
    # via sentry-sdk
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
element-api @ git+https://github.com/RUBclim/element-api
    # via -r requirements.in
fastapi==0.112.1
    # via
    #   -r requirements.in
    #   sentry-sdk
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.7
    # via anyio
kombu==5.3.7
    # via celery
numpy==2.1.0
    # via pandas
pandas==2.2.2
    # via element-api
prompt-toolkit==3.0.45
    # via -r requirements.in
psycopg-binary==3.1.19
    # via click-repl
psycopg[binary]==3.1.19
    # via psycopg
pydantic==2.8.2
    # via
    #   fastapi
    #   sqlmodel
pydantic-core==2.20.1
    # via pydantic
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   pandas
pytz==2024.1
    # via pandas
redis==5.0.4
    # via celery
sentry-sdk[fastapi]==2.4.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.32
    # via sqlmodel
sqlmodel==0.0.21
    # via -r requirements.in
starlette==0.38.2
    # via fastapi
typing-extensions==4.12.1
    # via
    #   fastapi
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzdata==2024.1
    # via
    #   celery
    #   pandas
urllib3==2.2.1
    # via sentry-sdk
uvicorn==0.30.6
    # via -r requirements.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
